+++ api/models/facebook/Feed.js
@@ -3,16 +3,16 @@ import moment from 'moment';
 module.exports = {
   attributes: {
     fullPicture: {
-      type: Sequelize.STRING,
+      type: Sequelize.TEXT,
     },
     name: {
       type: Sequelize.STRING,
     },
     description: {
-      type: Sequelize.STRING,
+      type: Sequelize.TEXT,
     },
     message: {
-      type: Sequelize.STRING,
+      type: Sequelize.TEXT,
     },
     story: {
       type: Sequelize.STRING,