+++ migration/v6/migrate.sql
@@ -1,3 +1,4 @@
+
 DROP TABLE Slogans;
 
 CREATE TABLE `Quotes` (
@@ -13,3 +14,11 @@ CREATE TABLE `Quotes` (
   KEY `img` (`img`),
   CONSTRAINT `quotes_ibfk_1` FOREIGN KEY (`img`) REFERENCES `Images` (`id`) ON DELETE SET NULL ON UPDATE CASCADE
 ) ENGINE=InnoDB AUTO_INCREMENT=2 DEFAULT CHARSET=utf8;
+
+
+ALTER TABLE RecipeOrders MODIFY COLUMN productionStatus enum("NEW","PENDING", "RECEIVED", "REQUESTED", "SUBMITTED", "PAID", "PROCESSING", "CANCELLED", "SHIPPED", "DELIVERED", "COMPLETED");
+ALTER TABLE EventOrders MODIFY COLUMN productionStatus enum("NEW","PENDING", "RECEIVED", "REQUESTED", "SUBMITTED", "PAID", "PROCESSING", "CANCELLED", "SHIPPED", "DELIVERED", "COMPLETED");
+UPDATE RecipeOrders JOIN Allpays ON Allpays.`RecipeOrderId` = RecipeOrders.`id` SET RecipeOrders.`productionStatus` = "PENDING" WHERE Allpays.PaymentType = 'aio';
+
+UPDATE Feeds SET publish = 0 WHERE type = 'status';
+ALTER TABLE Feeds MODIFY fullPicture text;