+++ migration/changelog.md
@@ -73,10 +73,26 @@
 ### RecipeOrder token 使用 UNIQUE
 ### EventOrder token 使用 UNIQUE
 
+### Feed 欄位變動
+* 修改 description 型態 STRING 修改為 TEXT
+* 修改 message 型態 STRING 修改為 TEXT
+
 ### ScentFeedback 欄位變動
 * 新增 feedbackCheck 確認 回饋內容 寫入 Scent 資料庫
 
 ## v6
+
 ### Slogans 表格移除
 
 ### 新增 Quotes 表格
+
+
+### RecipeOrder 欄位資料變動
+* productionStatus 增加 "PENDING"
+
+### EventOrder 欄位資料變動
+* productionStatus 增加 "PENDING"
+
+### Feed 欄位變動
+* 修改 fullPicture 欄位 型態 STRING 修改為 TEXT
+* type 欄位等於 'status'，將 publish欄位 修改為 false